refactor: move timing helpers out of pangu.ts
Extract once, debounce, and waitForVideosToLoad into
scheduling/timing.ts, unchanged, as part of splitting pangu.ts
per tmp/split-pangu-ts-handoff.md (steps 2 and 3, one file).
